Cleanup: deduplicate benchmarks
Changes:
- Delete old asv benchmarks
- Move python/benchmarks/rnns to top-level/rnns
- Add back python/benchmarks/benchmark_common.py and
  python/benchmarks/utils.py (I'm not sure how to preserve the blame on
  these, but I guess it's okay)
