--- python/benchmarks/lstm_variants/__init__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-from .lstm import *
-from .normalize import *
-from .container import *
