+++ python/benchmarks/mnist/README.md
