+++ python/benchmarks/mnist/main.py
