+++ python/benchmarks/mnist/requirements.txt
