+++ rnns/README.md
