+++ rnns/benchmarks/bnlstm.py
