+++ rnns/benchmarks/common.py
