+++ rnns/benchmarks/cudnn_lstm.py
