+++ rnns/benchmarks/lstm.py
