+++ rnns/benchmarks/lstm_variants/__init__.py
