+++ rnns/benchmarks/lstm_variants/container.py
