+++ rnns/benchmarks/lstm_variants/lstm.py
