+++ rnns/benchmarks/lstm_variants/normalize.py
