+++ rnns/benchmarks/lstm_variants_test.py
