+++ rnns/benchmarks/memnn.py
