+++ rnns/benchmarks/models/bnlstm.py
