+++ rnns/benchmarks/models/memnn.py
