+++ rnns/benchmarks/qrnn.py
