+++ rnns/benchmarks/sru_test.py
