+++ rnns/benchmarks/tensor.py
